% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcp1armBinary.R
\name{print.rcp1armBinary}
\alias{print.rcp1armBinary}
\title{Print Method for rcp1armBinary Objects}
\usage{
\method{print}{rcp1armBinary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rcp1armBinary"}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for rcp1armBinary Objects
}
