% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcp1armContinuous.R
\name{print.rcp1armContinuous}
\alias{print.rcp1armContinuous}
\title{Print Method for rcp1armContinuous Objects}
\usage{
\method{print}{rcp1armContinuous}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rcp1armContinuous"}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for rcp1armContinuous Objects
}
