% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcp1armCount.R
\name{print.rcp1armCount}
\alias{print.rcp1armCount}
\title{Print Method for rcp1armCount Objects}
\usage{
\method{print}{rcp1armCount}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rcp1armCount"}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for rcp1armCount Objects
}
