% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcp1armHazardRatio.R
\name{print.rcp1armHazardRatio}
\alias{print.rcp1armHazardRatio}
\title{Print Method for rcp1armHazardRatio Objects}
\usage{
\method{print}{rcp1armHazardRatio}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rcp1armHazardRatio"}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for rcp1armHazardRatio Objects
}
