% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcp1armMilestoneSurvival.R
\name{print.rcp1armMilestoneSurvival}
\alias{print.rcp1armMilestoneSurvival}
\title{Print Method for rcp1armMilestoneSurvival Objects}
\usage{
\method{print}{rcp1armMilestoneSurvival}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rcp1armMilestoneSurvival"}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for rcp1armMilestoneSurvival Objects
}
