% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcp1armRMST.R
\name{print.rcp1armRMST}
\alias{print.rcp1armRMST}
\title{Print Method for rcp1armRMST Objects}
\usage{
\method{print}{rcp1armRMST}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rcp1armRMST"}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Method for rcp1armRMST Objects
}
